+++ react/Summary.tsx
@@ -1,83 +1,53 @@
-import React from 'react'
-import { ExtensionPoint } from 'vtex.render-runtime'
+import React, {
+  createContext,
+  FunctionComponent,
+  ReactNode,
+  useContext,
+} from 'react'
 import { FormattedMessage, defineMessages } from 'react-intl'
-import { Button } from 'vtex.styleguide'
-import SummaryItem from './components/SummaryItem'
-
-import { OrderCouponProvider } from 'vtex.order-coupon/OrderCoupon'
 
 defineMessages({
   summary: {
     defaultMessage: 'Summary',
     id: 'store/checkout-summary.Summary',
   },
-  checkout: {
-    defaultMessage: 'Checkout',
-    id: 'store/checkout-summary.Checkout',
-  },
 })
 
-const minTotalizerValue = 0
-const shippingData = {
-  id: 'Shipping',
-  name: 'Total do Frete',
-  value: minTotalizerValue,
-  __typename: 'Totalizer',
+interface Context {
+  totalizers: any[]
+  total: number
 }
 
-const isShippingPresent = (totalizers: Totalizer[]) => {
-  let result = false
+const minTotalizerValue = 0
 
-  for (let totalizer of totalizers) {
-    if (totalizer.id === 'Shipping') {
-      result = true
-    }
+const SummaryContext = createContext<Context | undefined>(undefined)
+
+export const useSummary = () => {
+  const context = useContext(SummaryContext)
+  if (context === undefined) {
+    throw new Error('useSummary must be used within a SummaryProvider')
   }
 
-  return result
+  return context
 }
 
-const Summary: StorefrontFunctionComponent<SummaryProps> = ({
+const Summary: FunctionComponent<SummaryProps> = ({
+  children,
   totalizers,
   total,
 }) => {
-  if (!isShippingPresent(totalizers)) {
-    totalizers.push(shippingData)
-  }
-
   return (
-    <div className="c-on-base">
+    <SummaryContext.Provider value={{ totalizers: totalizers, total: total }}>
       <h5 className="t-heading-5 mt6 mb5">
         <FormattedMessage id="store/checkout-summary.Summary" />
       </h5>
-      <OrderCouponProvider>
-        <ExtensionPoint id="coupon" />
-      </OrderCouponProvider>
-
-      {totalizers.map(totalizer => (
-        <SummaryItem
-          key={totalizer.id}
-          label={totalizer.id}
-          name={totalizer.id === 'CustomTax' ? totalizer.name : ''}
-          value={(totalizer && totalizer.value) || minTotalizerValue}
-          large={false}
-        />
-      ))}
-
-      <SummaryItem
-        label="Total"
-        value={total ? total : minTotalizerValue}
-        large
-      />
-
-      <Button href="/checkout/#payment" variation="primary" size="large" block>
-        <FormattedMessage id="store/checkout-summary.Checkout" />
-      </Button>
-    </div>
+      <div className="c-on-base">{children}</div>
+    </SummaryContext.Provider>
   )
 }
 
 interface SummaryProps {
+  children: ReactNode
   title?: string
   intl: object
   totalizers: any[]