+++ react/SummaryCoupon.tsx
@@ -1,0 +1,3 @@
+import SummaryCoupon from './components/SummaryCoupon/SummaryCoupon'
+
+export default SummaryCoupon