+++ react/SummaryTotalizers.tsx
@@ -1,0 +1,3 @@
+import SummaryTotalizers from './components/SummaryTotalizers/SummaryTotalizers'
+
+export default SummaryTotalizers